refactor(blorktools): no main file
